// protocol/gen/ts/messages/drive_pb.ts
export class DriveTelemetry extends Message<DriveTelemetry> {
  /**
   * @generated from field: google.protobuf.Timestamp t = 1;
   */
  t?: Timestamp;

  /**
   * @generated from field: optional double body_vx_mps = 2;
   */
  bodyVxMps?: number;

  /**
   * @generated from field: optional double yaw_rate_radps = 3;
   */
  yawRateRadps?: number;

  /**
   * capture-time dual stamp
   *
   * @generated from field: waypoint.v1.Stamp stamp = 4;
   */
  stamp?: Stamp;

  constructor(data?: PartialMessage<DriveTelemetry>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "waypoint.v1.DriveTelemetry";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "t", kind: "message", T: Timestamp },
    { no: 2, name: "body_vx_mps", kind: "scalar", T: 1 /* ScalarType.DOUBLE */, opt: true },
    { no: 3, name: "yaw_rate_radps", kind: "scalar", T: 1 /* ScalarType.DOUBLE */, opt: true },
    { no: 4, name: "stamp", kind: "message", T: Stamp },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): DriveTelemetry {
    return new DriveTelemetry().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): DriveTelemetry {
    return new DriveTelemetry().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): DriveTelemetry {
    return new DriveTelemetry().fromJsonString(jsonString, options);
  }

  static equals(a: DriveTelemetry | PlainMessage<DriveTelemetry> | undefined, b: DriveTelemetry | PlainMessage<DriveTelemetry> | undefined): boolean {
    return proto3.util.equals(DriveTelemetry, a, b);
  }
}