// protocol/gen/ts/messages/image_pb.ts
export class ImageState extends Message<ImageState> {
  /**
   * @generated from field: string version = 1;
   */
  version = "";

  /**
   * "prod" or "dev"
   *
   * @generated from field: string variant = 2;
   */
  variant = "";

  /**
   * "A" or "B"
   *
   * @generated from field: string partition = 3;
   */
  partition = "";

  /**
   * @generated from field: int32 bootcount = 4;
   */
  bootcount = 0;

  /**
   * @generated from field: string built_at = 5;
   */
  builtAt = "";

  constructor(data?: PartialMessage<ImageState>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "waypoint.v1.ImageState";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "version", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "variant", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "partition", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 4, name: "bootcount", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
    { no: 5, name: "built_at", kind: "scalar", T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ImageState {
    return new ImageState().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ImageState {
    return new ImageState().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ImageState {
    return new ImageState().fromJsonString(jsonString, options);
  }

  static equals(a: ImageState | PlainMessage<ImageState> | undefined, b: ImageState | PlainMessage<ImageState> | undefined): boolean {
    return proto3.util.equals(ImageState, a, b);
  }
}