// protocol/gen/ts/messages/events_pb.ts
export class ModeEvent extends Message<ModeEvent> {
  /**
   * @generated from field: google.protobuf.Timestamp t = 1;
   */
  t?: Timestamp;

  /**
   * @generated from field: waypoint.v1.Mode from = 2;
   */
  from = Mode.UNSPECIFIED;

  /**
   * @generated from field: waypoint.v1.Mode to = 3;
   */
  to = Mode.UNSPECIFIED;

  /**
   * @generated from field: string reason = 4;
   */
  reason = "";

  constructor(data?: PartialMessage<ModeEvent>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "waypoint.v1.ModeEvent";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "t", kind: "message", T: Timestamp },
    { no: 2, name: "from", kind: "enum", T: proto3.getEnumType(Mode) },
    { no: 3, name: "to", kind: "enum", T: proto3.getEnumType(Mode) },
    { no: 4, name: "reason", kind: "scalar", T: 9 /* ScalarType.STRING */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): ModeEvent {
    return new ModeEvent().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): ModeEvent {
    return new ModeEvent().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): ModeEvent {
    return new ModeEvent().fromJsonString(jsonString, options);
  }

  static equals(a: ModeEvent | PlainMessage<ModeEvent> | undefined, b: ModeEvent | PlainMessage<ModeEvent> | undefined): boolean {
    return proto3.util.equals(ModeEvent, a, b);
  }
}