// protocol/gen/ts/messages/common_pb.ts
export enum Mode {
  /**
   * @generated from enum value: MODE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: MODE_MANUAL = 1;
   */
  MANUAL = 1,

  /**
   * @generated from enum value: MODE_SAFE = 2;
   */
  SAFE = 2,

  /**
   * @generated from enum value: MODE_AUTONOMOUS = 3;
   */
  AUTONOMOUS = 3,

  /**
   * @generated from enum value: MODE_ESTOP = 4;
   */
  ESTOP = 4,
}